- 운동 추가 안되는 문제 수정 - 운동 단위 텍스트 너비 수정

+++ HealthCare/src/main/java/com/example/healthcare/dataClass.kt
@@ -43,9 +43,9 @@ data class ExerciseType(
 
 data class ExerciseInfo(
     val exercise : String,
-    val weight : Int = 0,
-    val set : Int = 0,
-    val number : Int = 0
+    val weight : Float = 0f,
+    val set : Float = 0f,
+    val number : Float = 0f
 )
 
 data class ExerciseTypeList(
